checking if home-directory exists
git-svn-id: https://studio.atlassian.com/svn/AMPS/trunk@39045 2c54a935-e501-0410-bc05-97a93f6bca70

+++ maven-amps-plugin/src/main/java/com/atlassian/maven/plugins/amps/product/AbstractWebappProductHandler.java
@@ -224,7 +224,8 @@ public abstract class AbstractWebappProductHandler implements ProductHandler
     {
         final File srcDir = new File(project.getBasedir(), "src/test/resources/"+homeDirName);
         final File outputDir = new File(getBaseDirectory(), homeDirName);
-        FileUtils.copyDirectory(srcDir, outputDir);
+        if (srcDir.exists() && outputDir.exists())
+        	FileUtils.copyDirectory(srcDir, outputDir);
     }
